prediction-sentiment-pipeline/requirements.txt
google-cloud-bigquery
flask
google-cloud-language
google-cloud-logging
pandas
pandas-gbq